// backend/src/models/errorCodes.ts
export enum CUSTOM_ERROR_CODES {
    GENERAL_ERROR = 'GENERAL_ERROR',
    DATABASE_ERROR = 'DATABASE_ERROR',
    NOT_FOUND = 'NOT_FOUND',
    VALIDATION_FAILED = 'VALIDATION_FAILED',
    INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
    UNIQUE_CONSTRAINT_ERROR = 'UNIQUE_CONSTRAINT_ERROR',
    ACCOUNT_DOESNT_EXIST = 'ACCOUNT_DOESNT_EXIST',
    UNABLE_TO_SENT_REG_EMAIL = 'UNABLE_TO_SENT_REG_EMAIL',
    EMAIL_IN_USE = 'EMAIL_IN_USE',
    CONSTRAINT_FAILED = 'CONSTRAINT_FAILED',
}